pyproject.toml
[tool.poetry]
name = "blaise_dds"
version = "0.1.0"
description = ""
authors = ["Blaise 5 Team"]
packages = [
  { include = "blaise_dds" }
]

[tool.poetry.dependencies]
python = "^3.7"
google-auth = "^1.28.1"
google-oauth = "^1.0.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
black = "^20.8b1"
isort = "^5.7.0"
coverage = "^5.5"
flake8-cognitive-complexity = "^0.1.0"
pytest = "^6.2.3"
responses = "^0.13.2"
mypy = "^0.812"

[tool.isort]
multi_line_output = "3"
include_trailing_comma = "True"
line_length = "88"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
